# src/doc/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"_36_Enums.Role": {
				"type": "string",
				"enum": [
					"COACH",
					"ATHLETE_PROG",
					"ATHLETE_CO",
					"ATHLETE_FULL",
					"ADMIN"
				]
			},
			"Role": {
				"$ref": "#/components/schemas/_36_Enums.Role"
			},
			"CreateUserDto": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Tristan"
					},
					"surname": {
						"type": "string",
						"example": "Debord"
					},
					"age": {
						"type": "number",
						"format": "double",
						"example": 23
					},
					"weight": {
						"type": "number",
						"format": "double",
						"example": 85
					},
					"phone": {
						"type": "string",
						"example": "0601020304"
					},
					"username": {
						"type": "string",
						"example": "tristan@gmail.com"
					},
					"password": {
						"type": "string",
						"example": "1234"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"name",
					"surname",
					"age",
					"weight",
					"phone",
					"username",
					"password",
					"role"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "back",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"id": {
													"type": "number",
													"format": "double"
												},
												"password": {
													"type": "string"
												},
												"username": {
													"type": "string"
												},
												"imageUri": {
													"type": "string"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"phone": {
													"type": "string"
												},
												"weight": {
													"type": "number",
													"format": "double"
												},
												"age": {
													"type": "number",
													"format": "double"
												},
												"surname": {
													"type": "string"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"id",
												"password",
												"username",
												"imageUri",
												"role",
												"phone",
												"weight",
												"age",
												"surname",
												"name"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ALREADY-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"properties": {
													"stat": {
														"items": {
															"properties": {
																"deadlift": {
																	"type": "number",
																	"format": "double"
																},
																"bench": {
																	"type": "number",
																	"format": "double"
																},
																"squat": {
																	"type": "number",
																	"format": "double"
																},
																"userId": {
																	"type": "number",
																	"format": "double"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"deadlift",
																"bench",
																"squat",
																"userId",
																"id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"progUri": {
														"items": {
															"properties": {
																"userId": {
																	"type": "number",
																	"format": "double"
																},
																"fileUri": {
																	"type": "string"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																},
																"name": {
																	"type": "string"
																}
															},
															"required": [
																"userId",
																"fileUri",
																"id",
																"name"
															],
															"type": "object"
														},
														"type": "array"
													}
												},
												"required": [
													"stat",
													"progUri"
												],
												"type": "object"
											},
											{
												"properties": {
													"id": {
														"type": "number",
														"format": "double"
													},
													"username": {
														"type": "string"
													},
													"imageUri": {
														"type": "string"
													},
													"role": {
														"$ref": "#/components/schemas/_36_Enums.Role"
													},
													"phone": {
														"type": "string"
													},
													"weight": {
														"type": "number",
														"format": "double"
													},
													"age": {
														"type": "number",
														"format": "double"
													},
													"surname": {
														"type": "string"
													},
													"name": {
														"type": "string"
													}
												},
												"required": [
													"id",
													"username",
													"imageUri",
													"role",
													"phone",
													"weight",
													"age",
													"surname",
													"name"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/user/coach": {
			"post": {
				"operationId": "CreateCoach",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"id": {
													"type": "number",
													"format": "double"
												},
												"password": {
													"type": "string"
												},
												"username": {
													"type": "string"
												},
												"imageUri": {
													"type": "string"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"phone": {
													"type": "string"
												},
												"weight": {
													"type": "number",
													"format": "double"
												},
												"age": {
													"type": "number",
													"format": "double"
												},
												"surname": {
													"type": "string"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"id",
												"password",
												"username",
												"imageUri",
												"role",
												"phone",
												"weight",
												"age",
												"surname",
												"name"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ALREADY-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		},
		"/user/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"allOf": [
												{
													"properties": {
														"stat": {
															"items": {
																"properties": {
																	"deadlift": {
																		"type": "number",
																		"format": "double"
																	},
																	"bench": {
																		"type": "number",
																		"format": "double"
																	},
																	"squat": {
																		"type": "number",
																		"format": "double"
																	},
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	}
																},
																"required": [
																	"deadlift",
																	"bench",
																	"squat",
																	"userId",
																	"id"
																],
																"type": "object"
															},
															"type": "array"
														},
														"progUri": {
															"items": {
																"properties": {
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"fileUri": {
																		"type": "string"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"userId",
																	"fileUri",
																	"id",
																	"name"
																],
																"type": "object"
															},
															"type": "array"
														}
													},
													"required": [
														"stat",
														"progUri"
													],
													"type": "object"
												},
												{
													"properties": {
														"id": {
															"type": "number",
															"format": "double"
														},
														"username": {
															"type": "string"
														},
														"imageUri": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.Role"
														},
														"phone": {
															"type": "string"
														},
														"weight": {
															"type": "number",
															"format": "double"
														},
														"age": {
															"type": "number",
															"format": "double"
														},
														"surname": {
															"type": "string"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"id",
														"username",
														"imageUri",
														"role",
														"phone",
														"weight",
														"age",
														"surname",
														"name"
													],
													"type": "object"
												}
											]
										},
										{
											"type": "string",
											"enum": [
												"NOT-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"allOf": [
												{
													"properties": {
														"stat": {
															"items": {
																"properties": {
																	"deadlift": {
																		"type": "number",
																		"format": "double"
																	},
																	"bench": {
																		"type": "number",
																		"format": "double"
																	},
																	"squat": {
																		"type": "number",
																		"format": "double"
																	},
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	}
																},
																"required": [
																	"deadlift",
																	"bench",
																	"squat",
																	"userId",
																	"id"
																],
																"type": "object"
															},
															"type": "array"
														},
														"progUri": {
															"items": {
																"properties": {
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"fileUri": {
																		"type": "string"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"userId",
																	"fileUri",
																	"id",
																	"name"
																],
																"type": "object"
															},
															"type": "array"
														}
													},
													"required": [
														"stat",
														"progUri"
													],
													"type": "object"
												},
												{
													"properties": {
														"id": {
															"type": "number",
															"format": "double"
														},
														"username": {
															"type": "string"
														},
														"imageUri": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.Role"
														},
														"phone": {
															"type": "string"
														},
														"weight": {
															"type": "number",
															"format": "double"
														},
														"age": {
															"type": "number",
															"format": "double"
														},
														"surname": {
															"type": "string"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"id",
														"username",
														"imageUri",
														"role",
														"phone",
														"weight",
														"age",
														"surname",
														"name"
													],
													"type": "object"
												}
											]
										},
										{
											"type": "string",
											"enum": [
												"NOT-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"role": {
										"$ref": "#/components/schemas/Role"
									},
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									},
									"phone": {
										"type": "string"
									},
									"weight": {
										"type": "number",
										"format": "double"
									},
									"age": {
										"type": "number",
										"format": "double"
									},
									"surname": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"role",
									"password",
									"username",
									"phone",
									"weight",
									"age",
									"surname",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"id": {
													"type": "number",
													"format": "double"
												},
												"password": {
													"type": "string"
												},
												"username": {
													"type": "string"
												},
												"imageUri": {
													"type": "string"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"phone": {
													"type": "string"
												},
												"weight": {
													"type": "number",
													"format": "double"
												},
												"age": {
													"type": "number",
													"format": "double"
												},
												"surname": {
													"type": "string"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"id",
												"password",
												"username",
												"imageUri",
												"role",
												"phone",
												"weight",
												"age",
												"surname",
												"name"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"NOT-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/get-coach": {
			"get": {
				"operationId": "FindAllCoach",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"properties": {
													"stat": {
														"items": {
															"properties": {
																"deadlift": {
																	"type": "number",
																	"format": "double"
																},
																"bench": {
																	"type": "number",
																	"format": "double"
																},
																"squat": {
																	"type": "number",
																	"format": "double"
																},
																"userId": {
																	"type": "number",
																	"format": "double"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"deadlift",
																"bench",
																"squat",
																"userId",
																"id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"progUri": {
														"items": {
															"properties": {
																"userId": {
																	"type": "number",
																	"format": "double"
																},
																"fileUri": {
																	"type": "string"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																},
																"name": {
																	"type": "string"
																}
															},
															"required": [
																"userId",
																"fileUri",
																"id",
																"name"
															],
															"type": "object"
														},
														"type": "array"
													}
												},
												"required": [
													"stat",
													"progUri"
												],
												"type": "object"
											},
											{
												"properties": {
													"id": {
														"type": "number",
														"format": "double"
													},
													"username": {
														"type": "string"
													},
													"imageUri": {
														"type": "string"
													},
													"role": {
														"$ref": "#/components/schemas/_36_Enums.Role"
													},
													"phone": {
														"type": "string"
													},
													"weight": {
														"type": "number",
														"format": "double"
													},
													"age": {
														"type": "number",
														"format": "double"
													},
													"surname": {
														"type": "string"
													},
													"name": {
														"type": "string"
													}
												},
												"required": [
													"id",
													"username",
													"imageUri",
													"role",
													"phone",
													"weight",
													"age",
													"surname",
													"name"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/user/{id}/profile-image": {
			"put": {
				"operationId": "ImageUpdate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"stat": {
													"items": {
														"properties": {
															"deadlift": {
																"type": "number",
																"format": "double"
															},
															"bench": {
																"type": "number",
																"format": "double"
															},
															"squat": {
																"type": "number",
																"format": "double"
															},
															"userId": {
																"type": "number",
																"format": "double"
															},
															"id": {
																"type": "number",
																"format": "double"
															}
														},
														"required": [
															"deadlift",
															"bench",
															"squat",
															"userId",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												},
												"progUri": {
													"items": {
														"properties": {
															"userId": {
																"type": "number",
																"format": "double"
															},
															"fileUri": {
																"type": "string"
															},
															"id": {
																"type": "number",
																"format": "double"
															},
															"name": {
																"type": "string"
															}
														},
														"required": [
															"userId",
															"fileUri",
															"id",
															"name"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"stat",
												"progUri"
											],
											"type": "object"
										},
										{
											"properties": {
												"id": {
													"type": "number",
													"format": "double"
												},
												"username": {
													"type": "string"
												},
												"imageUri": {
													"type": "string"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"phone": {
													"type": "string"
												},
												"weight": {
													"type": "number",
													"format": "double"
												},
												"age": {
													"type": "number",
													"format": "double"
												},
												"surname": {
													"type": "string"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"id",
												"username",
												"imageUri",
												"role",
												"phone",
												"weight",
												"age",
												"surname",
												"name"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"profileImage": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"profileImage"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "ResetProfileImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"stat": {
													"items": {
														"properties": {
															"deadlift": {
																"type": "number",
																"format": "double"
															},
															"bench": {
																"type": "number",
																"format": "double"
															},
															"squat": {
																"type": "number",
																"format": "double"
															},
															"userId": {
																"type": "number",
																"format": "double"
															},
															"id": {
																"type": "number",
																"format": "double"
															}
														},
														"required": [
															"deadlift",
															"bench",
															"squat",
															"userId",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												},
												"progUri": {
													"items": {
														"properties": {
															"userId": {
																"type": "number",
																"format": "double"
															},
															"fileUri": {
																"type": "string"
															},
															"id": {
																"type": "number",
																"format": "double"
															},
															"name": {
																"type": "string"
															}
														},
														"required": [
															"userId",
															"fileUri",
															"id",
															"name"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"stat",
												"progUri"
											],
											"type": "object"
										},
										{
											"properties": {
												"id": {
													"type": "number",
													"format": "double"
												},
												"username": {
													"type": "string"
												},
												"imageUri": {
													"type": "string"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"phone": {
													"type": "string"
												},
												"weight": {
													"type": "number",
													"format": "double"
												},
												"age": {
													"type": "number",
													"format": "double"
												},
												"surname": {
													"type": "string"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"id",
												"username",
												"imageUri",
												"role",
												"phone",
												"weight",
												"age",
												"surname",
												"name"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}